Fix for two issues
Issue in Template class:
Replace only witespace characters instead of all non alfanumerical
characters. Else signs like ></= will all be replaced with _ and as a
result lose their distiction.

Issue in GdtTagLib:
Because the value wasn't encoded as HTML signs like < gave the browser
the wrong idea of an opening for a tag. This would result in lose of
displayed data.